all: fix compatibility with msgpack 3.0

+++ client_tools.go
@@ -12,7 +12,7 @@ type IntKey struct {
 
 func (k IntKey) EncodeMsgpack(enc *msgpack.Encoder) error {
 	enc.EncodeArrayLen(1)
-	enc.EncodeInt(k.I)
+	enc.EncodeInt(int64(k.I))
 	return nil
 }
 
@@ -24,7 +24,7 @@ type UintKey struct {
 
 func (k UintKey) EncodeMsgpack(enc *msgpack.Encoder) error {
 	enc.EncodeArrayLen(1)
-	enc.EncodeUint(k.I)
+	enc.EncodeUint(uint64(k.I))
 	return nil
 }
 
@@ -48,8 +48,8 @@ type IntIntKey struct {
 
 func (k IntIntKey) EncodeMsgpack(enc *msgpack.Encoder) error {
 	enc.EncodeArrayLen(2)
-	enc.EncodeInt(k.I1)
-	enc.EncodeInt(k.I2)
+	enc.EncodeInt(int64(k.I1))
+	enc.EncodeInt(int64(k.I2))
 	return nil
 }
 
@@ -63,7 +63,7 @@ type Op struct {
 func (o Op) EncodeMsgpack(enc *msgpack.Encoder) error {
 	enc.EncodeArrayLen(3)
 	enc.EncodeString(o.Op)
-	enc.EncodeInt(o.Field)
+	enc.EncodeInt(int64(o.Field))
 	return enc.Encode(o.Arg)
 }
 
@@ -78,9 +78,9 @@ type OpSplice struct {
 func (o OpSplice) EncodeMsgpack(enc *msgpack.Encoder) error {
 	enc.EncodeArrayLen(5)
 	enc.EncodeString(o.Op)
-	enc.EncodeInt(o.Field)
-	enc.EncodeInt(o.Pos)
-	enc.EncodeInt(o.Len)
+	enc.EncodeInt(int64(o.Field))
+	enc.EncodeInt(int64(o.Pos))
+	enc.EncodeInt(int64(o.Len))
 	enc.EncodeString(o.Replace)
 	return nil
 }
